ref: extract getEmotionRules test helper
