+++ tests/js/sentry-test/utils.tsx
@@ -1,5 +1,29 @@
 import MockDate from 'mockdate';
 
+/**
+ * Emotion injects its styles through the CSSOM (`insertRule`), so they are not
+ * reflected in `getComputedStyle` under jsdom. Read the generated rules for an
+ * element directly off `document.styleSheets` instead.
+ */
+export function getEmotionRules(element: HTMLElement): string[] {
+  const classes = element.className.split(' ').filter(cls => cls.startsWith('css-'));
+  const rules: string[] = [];
+  for (const sheet of Array.from(document.styleSheets)) {
+    let sheetRules: CSSRuleList;
+    try {
+      sheetRules = sheet.cssRules;
+    } catch {
+      continue;
+    }
+    for (const rule of Array.from(sheetRules)) {
+      if (classes.some(cls => rule.cssText.includes(cls))) {
+        rules.push(rule.cssText);
+      }
+    }
+  }
+  return rules;
+}
+
 // Taken from https://stackoverflow.com/a/56859650/1015027
 function findTextWithMarkup(contentNode: null | Element, textMatch: string | RegExp) {
   const hasText = (node: Element): boolean => {
